Split classes in actors.scala into their own files
