+++ riff-raff/app/deployment/actors/DeployMetricsActor.scala
@@ -1,0 +1,48 @@
+package deployment.actors
+
+import java.util.UUID
+
+import akka.actor._
+import akka.pattern.ask
+import akka.util.Timeout
+import conf.TaskMetrics
+import controllers.Logging
+import org.joda.time.DateTime
+
+import scala.concurrent.Await
+import scala.concurrent.duration._
+import scala.util.Try
+
+object DeployMetricsActor {
+  trait Message
+  case class TaskStart(deployId: UUID, taskId: String, queueTime: DateTime, startTime: DateTime) extends Message
+  case class TaskComplete(deployId: UUID, taskId: String, finishTime: DateTime) extends Message
+  case class TaskCountRequest() extends Message
+
+  lazy val system = ActorSystem("deploy-metrics")
+  lazy val deployMetricsProcessor = system.actorOf(Props(new DeployMetricsActor))
+  def runningTaskCount: Int = {
+    implicit val timeout = Timeout(100 milliseconds)
+    val count = deployMetricsProcessor ? TaskCountRequest() mapTo manifest[Int]
+    Try {
+      Await.result(count, timeout.duration)
+    } getOrElse 0
+  }
+}
+
+class DeployMetricsActor extends Actor with Logging {
+  var runningTasks = Map.empty[(UUID, String), DateTime]
+  import DeployMetricsActor._
+  def receive = {
+    case TaskStart(deployId, taskId, queueTime, startTime) =>
+      TaskMetrics.TaskStartLatency.recordTimeSpent(startTime.getMillis - queueTime.getMillis)
+      runningTasks += ((deployId, taskId) -> startTime)
+    case TaskComplete(deployId, taskId, finishTime) =>
+      runningTasks.get((deployId, taskId)).foreach { startTime =>
+        TaskMetrics.TaskTimer.recordTimeSpent(finishTime.getMillis - startTime.getMillis)
+      }
+      runningTasks -= (deployId -> taskId)
+    case TaskCountRequest() =>
+      sender ! runningTasks.size
+  }
+}
