+++ riff-raff/conf/logback.xml
@@ -24,7 +24,7 @@
     <logger name="com.amazonaws.request" level="WARN" />
 
     <logger name="deployment" level="DEBUG" />
-    <!--<logger name="deployment.DeployCoordinator" level="DEBUG" />-->
+    <!--<logger name="deployment.actors.DeployCoordinator" level="DEBUG" />-->
     <!--<logger name="deployment.TaskRunner" level="DEBUG" />-->
     <!--<logger name="ci" level="DEBUG" />-->
     <!--<logger name="utils.ScheduledAgent" level="DEBUG" />-->
